Move NotificationDispatcher base class to a modul dispatcher.py
